script_webgpu::dom

Module bindings

Modules

assert 🔒

callback 🔒

Base classes to work with IDL callbacks.

cell 🔒

A shareable mutable container for the DOM.

codegen 🔒

conversions 🔒

dom 🔒

domstring 🔒

error 🔒

inheritance 🔒

The Castable trait.

interface 🔒

Machinery to initialise interface prototype objects and interface objects.

interfaces 🔒

iterable 🔒

Implementation of iterable<...> and iterable<..., ...> WebIDL declarations.

like 🔒

Implementation of setlike<...> and maplike<..., ...> WebIDL declarations.

num 🔒

The Finite<T> struct.

principals 🔒

proxyhandler 🔒

Utilities for the implementation of JSAPI proxy handlers.

realms 🔒

record 🔒

The Record (open-ended dictionary) type.

refcounted 🔒

A generic, safe mechanism by which DOM objects can be pinned and transferred between threads (or intra-thread for asynchronous events). Akin to Gecko’s nsMainThreadPtrHandle, this uses thread-safe reference counting and ensures that the actual SpiderMonkey GC integration occurs on the script thread via weak refcounts. Ownership of a Trusted<T> object means the DOM object of type T to which it points remains alive. Any other behaviour is undefined. To guarantee the lifetime of a DOM object when performing asynchronous operations, obtain a Trusted<T> from that object and pass it along with each operation. A usable pointer to the original DOM object can be obtained on the script thread from a Trusted<T> via the root method.

reflector 🔒

root 🔒

script_runtime 🔒

settings_stack 🔒

str 🔒

The ByteString struct.

structuredclone 🔒

trace 🔒

utils 🔒

weakref 🔒

Weak-referenceable JS-managed DOM objects.

wrap 🔒

Macros

cformat 🔒

Creates a CString using interpolation of runtime expressions. Basically a format! that produces a CString.

match_domstring_ascii 🔒

Use this to match &str against lazydomstring efficiently. You are only allowed to match ascii strings otherwise this macro will lead to wrong results.

match_domstring_ascii_inner 🔒

Traits

DomTypes 🔒