script_webgpu::dom::bindings::cell

Struct DomRefCell

pub struct DomRefCell<T> {
    value: RefCell<T>,
}

A mutable field in the DOM.

This extends the API of std::cell::RefCell to allow unsafe access in certain situations, with dynamic checking in debug builds.

Fields
value: RefCell<T>
Implementations

impl<T> DomRefCell<T>

pub unsafe fn borrow_for_layout(&self) -> &T

Return a reference to the contents. For use in layout only.

Safety

Unlike RefCell::borrow, this method is unsafe because it does not return a Ref, thus leaving the borrow flag untouched. Mutably borrowing the RefCell while the reference returned by this method is alive is undefined behaviour.

Panics

Panics if this is called from anywhere other than the layout thread

Panics if the value is currently mutably borrowed.

pub unsafe fn borrow_for_script_deallocation(&self) -> &mut T

Borrow the contents for the purpose of script deallocation.

Safety

Unlike RefCell::borrow, this method is unsafe because it does not return a Ref, thus leaving the borrow flag untouched. Mutably borrowing the RefCell while the reference returned by this method is alive is undefined behaviour.

Panics

Panics if this is called from anywhere other than the script thread.

pub unsafe fn borrow_mut_for_layout(&self) -> &mut T

Mutably borrow a cell for layout. Ideally this would use RefCell::try_borrow_mut_unguarded but that doesn’t exist yet.

Safety

Unlike RefCell::borrow, this method is unsafe because it does not return a Ref, thus leaving the borrow flag untouched. Mutably borrowing the RefCell while the reference returned by this method is alive is undefined behaviour.

Panics

Panics if this is called from anywhere other than the layout thread.

pub fn safe_borrow_mut<'a, 'no_cx, 'r>( &'a self, _no_gc: &'no_cx NoGC, ) -> RefMut<'r, T>
where 'a: 'r, 'no_cx: 'r,

Mutably borrows the wrapped value.

The borrow lasts until the returned RefMut exits scope. The value cannot be borrowed while this borrow is active.

By passing a &NoGC we statically prevent GC from being run while the borrow is active, to prevent panic when tracing (which calls borrow).

Example

In simple cases one can use NoGC to statically ensure no GC can happen in the whole DOM method:

use js::context::{JSContext, NoGC};
use script_bindings::cell::DomRefCell;
fn DomMethod(no_gc: &NoGC, cell: &DomRefCell<usize>) {
    let mut mutably_borrowed = cell.safe_borrow_mut(no_gc);
}

But in more complex cases, method might trigger a GC, and thus require a &mut JSContext. In that case &JSContext can be used in place of NoGC, which will make RefMut bounded to the lifetime of the &JSContext and thus prevent any GC from happening while it is alive.

use js::context::{JSContext, NoGC};
use script_bindings::cell::DomRefCell;
fn GC(cx: &mut JSContext) {}

fn DomMethod(cell: &DomRefCell<usize>, cx: &mut JSContext) {
    {
        let mut mutably_borrowed = cell.safe_borrow_mut(cx);
        // do something with mutably_borrowed

        // only &JSContext is available here
    } // mutably_borrowed goes out of scope here
    // so one can now use &mut JSContext
    GC(cx);
}

use js::context::{JSContext, NoGC};
use script_bindings::cell::DomRefCell;
fn GC(cx: &mut JSContext) {}

fn DomMethod(cell: &DomRefCell<usize>, cx: &mut JSContext) {
    {
        let mut mutably_borrowed = cell.safe_borrow_mut(cx);
        // do something with mutably_borrowed

        // here one cannot use anything that might trigger a GC
        // as that would require &mut JSContext
        // but there is already existing &JSContext bounded at RefMut
        GC(cx);
    } // mutably_borrowed goes out of scope here
}

Panics

Panics if the value is currently borrowed.

pub fn safe_try_borrow_mut<'a, 'no_cx, 'r>( &'a self, _no_gc: &'no_cx NoGC, ) -> Result<RefMut<'r, T>, BorrowMutError>
where 'a: 'r, 'no_cx: 'r,

Mutably borrows the wrapped value.

The borrow lasts until the returned RefMut exits scope. The value cannot be borrowed while this borrow is active.

By passing a &NoGC we statically prevent GC from being run while the borrow is active, to prevent panic when tracing (which calls borrow).

Returns None if the value is currently borrowed.

impl<T> DomRefCell<T>

pub fn new(value: T) -> DomRefCell<T>

Create a new DomRefCell containing value.

pub fn borrow(&self) -> Ref<'_, T>

Immutably borrows the wrapped value.

The borrow lasts until the returned Ref exits scope. Multiple immutable borrows can be taken out at the same time.

Panics

Panics if the value is currently mutably borrowed. Panics if this is called from anywhere other than the script thread. Use borrow_for_layout if the borrowed data might used during layout.

pub fn borrow_mut(&self) -> RefMut<'_, T>

Mutably borrows the wrapped value.

The borrow lasts until the returned RefMut exits scope. The value cannot be borrowed while this borrow is active.

Panics

Panics if the value is currently borrowed. Panics if this is called from anywhere other than the script thread.

pub fn try_borrow(&self) -> Result<Ref<'_, T>, BorrowError>

Attempts to immutably borrow the wrapped value.

The borrow lasts until the returned Ref exits scope. Multiple immutable borrows can be taken out at the same time.

Returns None if the value is currently mutably borrowed.

Panics

Panics if this is called off the script thread.

pub fn try_borrow_mut(&self) -> Result<RefMut<'_, T>, BorrowMutError>

Mutably borrows the wrapped value.

The borrow lasts until the returned RefMut exits scope. The value cannot be borrowed while this borrow is active.

Returns None if the value is currently borrowed.

Panics

Panics if this is called off the script thread.

impl<T> DomRefCell<T>
where T: Default,

pub fn take(&self) -> T

Takes the wrapped value, leaving Default::default() in its place.

Panics

Panics if the value is currently borrowed.