script_webgpu::dom::bindings

Module domstring

Structs

BytesView

DOMString

This type corresponds to the DOMString type in WebIDL.

StringView

A reference to a Rust str of UTF-8 encoded bytes, used to get a Rust string from a DOMString.

Enums

DOMStringErrorType

EncodedBytes

A type representing the underlying encoded bytes of a DOMString.

Functions

parse_floating_point_number

https://html.spec.whatwg.org/multipage/#rules-for-parsing-floating-point-number-values