script_webgpu::dom::bindings::like

Trait Setlike

pub trait Setlike {
    type Key: ToJSValConvertible + Clone;

    // Required methods
    fn get_index(&self, cx: &mut JSContext, index: u32) -> Option<Self::Key>;
    fn size(&self, cx: &mut JSContext) -> u32;
    fn add(&self, cx: &mut JSContext, key: Self::Key);
    fn has(&self, cx: &mut JSContext, key: Self::Key) -> bool;
    fn clear(&self, cx: &mut JSContext);
    fn delete(&self, cx: &mut JSContext, key: Self::Key) -> bool;
}

Every Setlike dom_struct must implement this to provide access to underlying storage so codegen can automatically generate all setlike methods

In case you use a type that implements Setlike as underlying storage it’s recommended to use setlike macro.

Required Associated Types

type Key: ToJSValConvertible + Clone

The type of the key of the set.

Required Methods

fn get_index(&self, cx: &mut JSContext, index: u32) -> Option<Self::Key>

fn size(&self, cx: &mut JSContext) -> u32

fn add(&self, cx: &mut JSContext, key: Self::Key)

fn has(&self, cx: &mut JSContext, key: Self::Key) -> bool

fn clear(&self, cx: &mut JSContext)

fn delete(&self, cx: &mut JSContext, key: Self::Key) -> bool