script_webgpu::dom::bindings

Macro match_domstring_ascii

macro_rules! match_domstring_ascii {
    ($input:expr, $($tail:tt)*) => { ... };
}

Use this to match &str against lazydomstring efficiently. You are only allowed to match ascii strings otherwise this macro will lead to wrong results.

let s = DOMString::from("test");
let value = match_domstring!(s,
"test1" => 1,
"test2" => 2,
"test" => 3,
_ => 4,
);
assert_eq!(value, 3);

The RefCell inside DOMString is borrowed for the duration of the match, so the string cannot be accessed again inside a match arm.