script_webgpu::dom::bindings

Module proxyhandler

Utilities for the implementation of JSAPI proxy handlers.

Structs

CrossOriginProperties

Property and method specs that correspond to the elements of CrossOriginProperties(O).

Functions

cross_origin_get

Implementation of CrossOriginGet.

cross_origin_get_own_property_helper

Implementation of CrossOriginGetOwnPropertyHelper.

cross_origin_own_property_keys

Implementation of CrossOriginOwnPropertyKeys.

cross_origin_property_fallback

Implementation of CrossOriginPropertyFallback.

cross_origin_set⚠

Implementation of CrossOriginSet.

init

Initialize the infrastructure for DOM proxy objects.

is_extensible⚠

Reports whether the object is Extensible

is_platform_object_same_origin

https://html.spec.whatwg.org/multipage/#isplatformobjectsameorigin-(-o-)

maybe_cross_origin_get_prototype

Implementation of [[GetPrototypeOf]] for Location.

maybe_cross_origin_get_prototype_if_ordinary_rawcx⚠

Safety

maybe_cross_origin_set_prototype_rawcx⚠

Implementation of [[SetPrototypeOf]] for Location and WindowProxy.

prevent_extensions⚠

Controls whether the Extensible bit can be changed

report_cross_origin_denial

Report a cross-origin denial for a property, Always returns false, so it can be used as return report_cross_origin_denial(...);.

set_property_descriptor

Set the property descriptor’s object to obj and set it to enumerable, and writable if readonly is true.