script_webgpu::dom::bindings

Module reflector

Structs

AssociatedMemory

Reflector

A struct to store a reference to the reflector of a DOM object.

Traits

AssociatedMemorySize

DomGlobalGeneric

DomObject

A trait to provide access to the Reflector for a DOM object.

DomObjectIteratorWrap

A trait to provide a function pointer to wrap function for DOM iterator interfaces.

DomObjectWrap

A trait to provide a function pointer to wrap function for DOM objects.

MutDomObject

A trait to initialize the Reflector for a DOM object.

WeakReferenceableDomObjectWrap

A trait to provide a function pointer to wrap function for DOM objects.

Functions

reflect_dom_object

Create the reflector for a new DOM object and yield ownership to the reflector.

reflect_dom_object_with_cx

Create the reflector for a new DOM object and yield ownership to the reflector. Deprecated, use reflect_dom_object instead.

reflect_dom_object_with_proto

reflect_dom_object_with_proto_and_wrap

Create the reflector for a new DOM object and yield ownership to the reflector.

reflect_dom_object_with_wrap

Create the reflector for a new DOM object and yield ownership to the reflector.

reflect_weak_referenceable_dom_object

Create the reflector for a new DOM object and yield ownership to the reflector.

reflect_weak_referenceable_dom_object_with_cx_and_wrap

Create the reflector for a new DOM object and yield ownership to the reflector.

reflect_weak_referenceable_dom_object_with_proto