script_webgpu::dom::bindings::reflector

Trait MutDomObject

pub trait MutDomObject: DomObject {
    // Required methods
    unsafe fn init_reflector<D>(&self, obj: *mut JSObject);
    unsafe fn init_reflector_without_associated_memory(
        &self,
        obj: *mut JSObject,
    );
}

A trait to initialize the Reflector for a DOM object.

Required Methods

unsafe fn init_reflector<D>(&self, obj: *mut JSObject)

Initializes the Reflector

Safety

The provided JSObject pointer must point to a valid JSObject. The provided JSObject pointer must not be allocated in the nursery.

unsafe fn init_reflector_without_associated_memory(&self, obj: *mut JSObject)

Initializes the Reflector without recording any associated memory usage.

Safety

The provided JSObject pointer must point to a valid JSObject.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.