script_webgpu::dom::bindings

Module str

The ByteString struct.

Structs

ByteString

Encapsulates the IDL ByteString type.

DOMString

This type corresponds to the DOMString type in WebIDL.

USVString

A string that is constructed from a UCS-2 buffer by replacing invalid code points with the replacement character.

Functions

is_token

Returns whether s is a token, as defined by RFC 2616.

serialize_jsval_to_json_utf8

Because this converts to a DOMString it becomes UTF-8 encoded which is closer to the spec definition of https://infra.spec.whatwg.org/#serialize-a-javascript-value-to-json-bytes but we generally do not operate on anything that is truly a WTF-16 string.