script_webgpu::dom::bindings::str

Struct DOMString

#[repr(transparent)]
pub struct DOMString(RefCell<DOMStringType>);

This type corresponds to the DOMString type in WebIDL.

Conceptually, a DOMString has the same value space as a JavaScript String, i.e., an array of 16-bit code units representing UTF-16, potentially with unpaired surrogates present (also sometimes called WTF-16).

However, Rust Strings are guaranteed to be valid UTF-8, and as such have a smaller value space than WTF-16 (i.e., some JavaScript String values can not be represented as a Rust String). This introduces the question of what to do with values being passed from JavaScript to Rust that contain unpaired surrogates.

The hypothesis is that it does not matter much how exactly those values are transformed, because passing unpaired surrogates into the DOM is very rare. Instead Servo will replace the unpaired surrogate by a U+FFFD replacement character.

Currently, the lack of crash reports about this issue provides some evidence to support the hypothesis. This evidence will hopefully be used to convince other browser vendors that it would be safe to replace unpaired surrogates at the boundary between JavaScript and native code. (This would unify the DOMString and USVString types, both in the WebIDL standard and in Servo.)

This string class will keep either the Reference to the mozjs object alive or will have an internal rust string. We currently default to doing most of the string operation on the rust side. You should use str() to get the Rust string (represented by a StringView which you can deref to a &str). You should assume that this conversion is expensive. For now, you should assume that all the functions incur this conversion cost.

Tuple Fields
0: RefCell<DOMStringType>
Implementations

impl DOMString

pub fn new() -> DOMString

Creates a new DOMString.

pub fn from_js_string( cx: &mut JSContext, value: Handle<'_, Value>, ) -> Result<DOMString, DOMStringErrorType>

Creates the string from js. If the string can be encoded in latin1, just take the reference to the JSString. Otherwise do the conversion to utf8 now.

pub fn from_static(s: &'static str) -> DOMString

Creates a DOMString from a &'static str reference. More efficient than allocating the string.

pub fn str(&self) -> StringView<'_>

Returns the underlying rust string.

pub fn encoded_bytes(&self) -> EncodedBytes<'_>

Return the EncodedBytes of this DOMString. This returns the original encoded bytes of the string without doing any conversions.

pub fn clear(&mut self)

pub fn is_empty(&self) -> bool

pub fn len(&self) -> usize

The length of this string in UTF-8 code units, each one being one byte in size.

Note: This is different than the number of Unicode characters (or code points). A character may require multiple UTF-8 code units.

pub fn len_utf8(&self) -> Utf8CodeUnits

The length of this string in UTF-8 code units, each one being one byte in size. This method is the same as DOMString::len, but the result is wrapped in a Utf8CodeUnits to be used in code that mixes different kinds of offsets.

Note: This is different than the number of Unicode characters (or code points). A character may require multiple UTF-8 code units.

pub fn len_utf16(&self) -> Utf16CodeUnits

The length of this string in UTF-16 code units, each one being one two bytes in size.

Note: This is different than the number of Unicode characters (or code points). A character may require multiple UTF-16 code units.

pub fn make_ascii_lowercase(&mut self)

This works the same as make_ascii_lowercase on std::string. This means that any character in [A-Z] will be transformed to lower case but other characters stay the same (either ASCII or not ASCII).

pub fn push_str(&mut self, string_to_push: &str)

pub fn strip_leading_and_trailing_ascii_whitespace(&mut self)

https://infra.spec.whatwg.org/#strip-leading-and-trailing-ascii-whitespace

pub fn is_valid_floating_point_number_string(&self) -> bool

https://html.spec.whatwg.org/multipage/#valid-floating-point-number

pub fn parse<T>(&self) -> Result<T, <T as FromStr>::Err>
where T: FromStr,

pub fn parse_floating_point_number(&self) -> Option<f64>

https://html.spec.whatwg.org/multipage/#rules-for-parsing-floating-point-number-values

pub fn set_best_representation_of_the_floating_point_number(&mut self)

https://html.spec.whatwg.org/multipage/#best-representation-of-the-number-as-a-floating-point-number

pub fn to_lowercase(&self) -> String

pub fn to_uppercase(&self) -> String

pub fn strip_newlines(&mut self)

pub fn normalize_newlines(&mut self)

Normalize newlines according to https://infra.spec.whatwg.org/#normalize-newlines.

pub fn replace(self, needle: &str, replace_char: &str) -> DOMString

pub fn starts_with(&self, c: char) -> bool

Pattern is not yet stable in rust, hence, we need different methods for str and char

pub fn starts_with_str(&self, needle: &str) -> bool

pub fn ends_with_str(&self, needle: &str) -> bool

pub fn contains(&self, needle: &str) -> bool

pub fn eq_ignore_ascii_case(&self, other: &str) -> bool

Returns whether this DOMString is an ASCII case-insensitive match for other, without allocating and copying temporaries.

https://infra.spec.whatwg.org/#ascii-case-insensitive

pub fn to_ascii_lowercase(&self) -> String

pub fn contains_tab_or_newline(&self) -> bool

https://infra.spec.whatwg.org/#ascii-tab-or-newline

pub fn contains_html_space_characters(&self) -> bool

https://infra.spec.whatwg.org/#ascii-whitespace

pub fn as_bytes(&self) -> BytesView<'_>

This returns the string in utf8 bytes, i.e., [u8] encoded with utf8.

pub fn is_ascii_lowercase(&self) -> bool

Tests if there are only ascii lowercase characters. Does not include special characters.

pub fn is_ascii(&self) -> bool

Is the string only ascii characters

pub fn is_valid_for_cookie(&self) -> bool

Returns true if the slice only contains bytes that are safe to use in cookie strings. https://www.ietf.org/archive/id/draft-ietf-httpbis-rfc6265bis-15.html#section-5.6-6 Not using ServoCookie::is_valid_name_or_value to prevent dependency on the net crate.

pub fn normalize_crlf(&self) -> String

Newline replacement routine as described in step 1 of the multipart/form-data encoding algorithm and many steps of application/x-www-form-urlencoded. e.g. https://html.spec.whatwg.org/multipage/#convert-to-a-list-of-name-value-pairs

Replace every occurrence of U+000D (CR) not followed by U+000A (LF), and every occurrence of U+000A (LF) not preceded by U+000D (CR), in entry’s name, by a string consisting of a U+000D (CR) and U+000A (LF).