script_webgpu::dom::bindings

Module utils

Modules

call_policies

Structs

CallPolicyInfo

Controls various details of an IDL operation, such as whether a [LegacyLenientThis] attribute is specified and preconditions that affect the outcome of the “perform a security check” steps.

DOMClass

The struct that holds inheritance information for DOM object reflectors.

DOMJSClass

The JSClass used for DOM object reflectors.

Traits

CallPolicy

Functions

define_dictionary_property

Define an own enumerable data property with name property on object. Returns Err(()) on JSAPI failure, or null object, and Ok(()) otherwise.

get_array_index_from_id

Get an array index from the given jsid. Returns None if the given jsid is not an integer.

has_own_property

Checks whether object has an own property named property. Returns Err(()) on JSAPI failure (there is a pending exception), and Ok(false) for null objects or when the property is not own.

has_property_on_prototype

Computes whether proxy has a property id on its prototype and stores the result in found.

set_dictionary_property

Set the property with name property from object. Returns Err(()) on JSAPI failure, or null object, and Ok(()) otherwise

to_frozen_array

Returns a JSVal representing the frozen JavaScript array

Type Aliases

ProtoOrIfaceArray

An array of *mut JSObject of size PROTO_OR_IFACE_LENGTH.