script_webgpu

Module gpuconvert

Traits

WebGPUConvert

A version of the Into<T> trait from the standard library that can be used to convert between two types that are not defined in the script crate. This is intended to be used on dict/enum types generated from WebIDL once those types are moved out of the script crate. Only for WebGPU.

WebGPUTryConvert

A version of the TryInto<T> trait from the standard library that can be used to convert between two types that are not defined in the script crate. This is intended to be used on dict/enum types generated from WebIDL once those types are moved out of the script crate. Only for WebGPU.

Functions

convert_bind_group_entry 🔒

convert_bind_group_layout_entry 🔒

convert_load_op

convert_texture_descriptor

convert_texture_for_wgpu_with_cx