script_webgpu::gpuconvert

Trait WebGPUConvert

pub trait WebGPUConvert<T> {
    // Required method
    fn convert(self) -> T;
}

A version of the Into<T> trait from the standard library that can be used to convert between two types that are not defined in the script crate. This is intended to be used on dict/enum types generated from WebIDL once those types are moved out of the script crate. Only for WebGPU.

Required Methods

fn convert(self) -> T