script_webgpu::gpuconvert

Trait WebGPUTryConvert

pub trait WebGPUTryConvert<T> {
    type Error;

    // Required method
    fn try_convert(self) -> Result<T, Self::Error>;
}

A version of the TryInto<T> trait from the standard library that can be used to convert between two types that are not defined in the script crate. This is intended to be used on dict/enum types generated from WebIDL once those types are moved out of the script crate. Only for WebGPU.

Required Associated Types

type Error

Required Methods

fn try_convert(self) -> Result<T, Self::Error>